Refresh SSO AccessToken if first attempt errors out
For some reason our SSO AccessTokens are being marked as invalid
by AWS SSO API and so the sso.GetAccounts() and sso.GetRoles()
calls should try force reauthenticating once to AWS SSO in order to get a
new AccessToken for these calls.

Fixes: #279
