+++ .github/workflows/codeql-analysis.yml
@@ -40,8 +40,9 @@ jobs:
 
     # If this run was triggered by a pull request event, then checkout
     # the head of the pull request instead of the merge commit.
-    - run: git checkout HEAD^2
-      if: ${{ github.event_name == 'pull_request' }}
+    # Note: no longer recommended!
+    #    - run: git checkout HEAD^2
+    #  if: ${{ github.event_name == 'pull_request' }}
 
     # Initializes the CodeQL tools for scanning.
     - name: Initialize CodeQL
